354 - `SearchTriggerLogicSpec` already existed! Just extracted out from another spec
